requirements.txt
# File: requirements.txt
notion-client
python-dotenv
requests
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib